+++ src/command_modules/azure-cli-interactive/azclishell/az_completer.py
@@ -4,6 +4,7 @@
 # --------------------------------------------------------------------------------------------
 
 from __future__ import absolute_import, division, print_function, unicode_literals
+import sys
 
 from prompt_toolkit.completion import Completer, Completion
 
@@ -17,6 +18,12 @@ from azure.cli.core.parser import AzCliCommandParser
 from azure.cli.core.util import CLIError
 
 SELECT_SYMBOL = azclishell.configuration.SELECT_SYMBOL
+
+BLACKLISTED_COMPLETIONS = ['interactive']
+
+
+def error_pass(_, message):  # pylint: disable=unused-argument
+    return
 
 
 def dynamic_param_logic(text):
@@ -75,7 +82,11 @@ def sort_completions(gen):
             priority = ' '  # a space has the lowest ordinance
         return priority + val.text
 
-    return sorted(list(gen), key=_get_weight)
+    completions = []
+    for comp in gen:
+        if comp.text not in BLACKLISTED_COMPLETIONS:
+            completions.append(comp)
+    return sorted(completions, key=_get_weight)
 
 
 # pylint: disable=too-many-instance-attributes
@@ -109,6 +120,7 @@ class AzCompleter(Completer):
         self.output_options = commands.output_options if global_params else []
         self.global_param_descriptions = commands.global_param_descriptions if global_params else []
 
+        AzCliCommandParser.error = error_pass  # mutes the parsing
         self.global_parser = AzCliCommandParser(add_help=False)
         self.global_parser.add_argument_group('global', 'Global Arguments')
         self.parser = AzCliCommandParser(parents=[self.global_parser])
@@ -199,7 +211,6 @@ class AzCompleter(Completer):
                             for comp in gen_dyn_completion(
                                     comp, started_param, prefix, text):
                                 yield comp
-
                     except TypeError:
                         try:
                             for comp in self.cmdtab[self.curr_command].\