+++ src/command_modules/azure-cli-rdbms/azure/cli/command_modules/rdbms/__init__.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-import azure.cli.command_modules.rdbms.help  # pylint: disable=unused-import
+import azure.cli.command_modules.rdbms._help  # pylint: disable=unused-import
 
 __all__ = ['load_params', 'load_commands']