+++ src/command_modules/azure-cli-rdbms/tests/test_rdbms_commands.py
@@ -364,10 +364,15 @@ class ProxyResourcesMgmtScenarioTest(ScenarioTest):
                          JMESPathCheck('value', new_value)])
 
         # test list log files
-        result = self.cmd('{} server-logs list -g {} -s {}'
+        result = self.cmd('{} server-logs list -g {} -s {} --file-last-written 43800'  # ensure recording good for at least 5 years!
                           .format(database_engine, rg, server),
                           checks=[
                               JMESPathCheck('length(@)', 1),
                               JMESPathCheck('type(@)', 'array')]).get_output_in_json()
 
         self.assertIsNotNone(result[0]['name'])
+
+
+if __name__ == '__main__':
+    import unittest
+    unittest.main()